refactor: 리뷰 반영
- mock을 용도, 서비스에 따라 분리
- socket.io-client devdeps로 이동
- logger 레벨 제대로 설정 debug -> warn
