+++ backend/signal/mocks/events/logger.mock.ts
@@ -1,0 +1,7 @@
+export const loggerServiceMock = {
+  debug: jest.fn(),
+  info: jest.fn(),
+  warn: jest.fn(),
+  error: jest.fn(),
+  fatal: jest.fn(),
+};
