+++ backend/signal/package.json
@@ -31,7 +31,6 @@
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
     "socket.io": "^4.7.2",
-    "socket.io-client": "^4.7.3",
     "uuid": "^9.0.1",
     "winston": "^3.11.0"
   },
@@ -58,7 +57,8 @@
     "ts-loader": "^9.4.3",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.1.3"
+    "typescript": "^5.1.3",
+    "socket.io-client": "^4.7.3"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -78,6 +78,9 @@
     "testEnvironment": "node",
     "moduleNameMapper": {
       "^src/(.*)$": "<rootDir>/$1"
-    }
+    },
+    "modulePaths": [
+      "<rootDir>/.."
+    ]
   }
 }
